// build-test/src/types/request.d.ts
export declare namespace RequestParams {
  interface Login {
    name: string;
  }

  interface CreateProductParams {
    name: string;
    remark: string;
  }
  interface UpdateProductParams {
    name: string;
    remark: string;
    id: number;
  }
  interface GetProductList {
    name: string;
    // manager: string;
    pageIndex: number;
    pageSize: number;
  }
  interface UpdateStatus {
    id: number;
    status: number;
  }
  interface IterationParams {
    status: number;
    pageIndex: number;
    pageSize: number;
  }
  interface CreateDemand {
    collect_time?: string;
    name: string;
    product_id: number | string;
    create_by?: string;
    creator?: string;
    description: string;
    id?: number;
    status?: number;
    demand_status?: number;
    level: number;
    type: number;
    parent_id?: number;
    product_id: number;
    name: string;
    origin?: string;
    origin_path?: string;
    iteration_name?: string;
    child_list?: any;
    list?: any;
    origin_remark: string;
    product_module_id: number;
    product_module_name?: string;
    attachment: Record<string, any>;
  }
  interface GetUserDemandList {
    product_id: [];
  }
  interface GetDemandList {
    pageIndex: number;
    page_size: number;
    pageSize: number;
    status: number;
    product_id: number;
  }
  interface GetDemandListDetail {
    id: number;
  }
  interface PutDemand {
    id: number;
    product_id: number;
    name: string;
    child_list: {
      id: number;
      name: string;
    };
  }
// ...
  interface UpdatePlanList {
    id?: number | null;
    end_time: string;
    start_time: string;
    demand_ids: Array<number | string> | string;
    name: string;
    product_id?: number | null;
    create_by?: string;
    creator?: string;
    staff_no: string;
    staff_name: string;
    remark: string;
  }

  interface GetRelevanceDemandList {
    product_id: number | undefined;
    plan_id: number | undefined;
    is_bind: boolean | undefined;
  }
// ...
  interface GetPlan {
    create_by: Array<any>;
    name: string;
    product_id: Array<any>;
  }
// ...
}