// build-test/src/types/interface.ts
export interface RequestInter {
  index?: number;
  name: string | number;
  level?: number;
  type?: number;
  origin_path: string;
  description: string;
  estimate: string;
  target_value: string;
  id: number;
}

export interface ApprovalEditForm {
  title: string;
  attachment: DementApprovalAttachment[];
  product_id?: number;
  id?: number;
  demand_list: RequestInter[];
}

export interface DemandApprovalItem {
  create_by: string;
  create_time: string;
  creator: string;
  delete_time: string;
  demand_status: number;
  description: string;
  id: number;
  level: number;
  name: string;
  origin: string;
  origin_path: string;
  origin_remark: string;
  parent_id: number;
  product_id: number;
  product_module_id: number;
  type: number;
}

export interface DementApprovalAttachment {
  addr: string;
  size: string;
  title: string;
}